Code to render from data file. Based on Saveliy's rendering code
Basically gutted all the server and threading stuff and have a slower version that just reads from a file and shows it. File format is pure bytes with 10 bytes of 0xff as marker and then 2 bytes of the angle and then the data values. Integrates data conversion directly into the rendering program.